+++ src/repos/commands.ts
@@ -8,21 +8,38 @@ import {
   workspace
 } from "vscode";
 import { EXTENSION_NAME } from "../constants";
+import { store as globalStore } from "../store";
+import { getCurrentUser } from "../store/auth";
 import { withProgress } from "../utils";
+import { RepoFileSystemProvider } from "./fileSystem";
+import { Repository } from "./store";
 import {
   addRepoFile,
-  deleteRepoFile,
+  createBranch,
+  deleteBranch,
+  deleteRepository,
   listRepos,
   manageRepo,
+  rebaseBranch,
   refreshRepositories,
-  renameFile,
   unmanageRepo
 } from "./store/actions";
 import { RepositoryFileNode, RepositoryNode } from "./tree/nodes";
+import moment = require("moment");
 
-function getGitHubUrl(repo: string, filePath?: string) {
-  const suffix = filePath ? `/blob/master/${filePath}` : "";
+function getGitHubUrl(repo: string, branch: string, filePath?: string) {
+  const suffix = filePath
+    ? `/blob/${branch}/${filePath}`
+    : branch !== Repository.DEFAULT_BRANCH
+    ? `/tree/${branch}`
+    : "";
+
   return `https://github.com/${repo}${suffix}`;
+}
+
+function openRepoDocument(repo: string, file: string) {
+  const uri = RepoFileSystemProvider.getFileUri(repo, file);
+  window.showTextDocument(uri);
 }
 
 let repoPromise: Promise<any>;
@@ -37,7 +54,7 @@ export async function registerRepoCommands(context: ExtensionContext) {
       }
 
       quickPick.busy = true;
-      quickPick.placeholder = "";
+      quickPick.placeholder = "Loading your repositories...";
 
       repoPromise.then((repos) => {
         const items = repos.map((repo: any) => ({
@@ -56,8 +73,7 @@ export async function registerRepoCommands(context: ExtensionContext) {
         });
 
         quickPick.busy = false;
-        quickPick.placeholder =
-          "Select or specify the repo to manage (e.g. vsls-contrib/guestbook)";
+        quickPick.placeholder = "";
       });
 
       quickPick.onDidAccept(async () => {
@@ -79,7 +95,7 @@ export async function registerRepoCommands(context: ExtensionContext) {
         const nodes = multiSelectNodes || [targetNode];
 
         for (const node of nodes) {
-          await unmanageRepo(node.repo.name);
+          await unmanageRepo(node.repo.name, node.repo.branch);
         }
       }
     )
@@ -89,7 +105,7 @@ export async function registerRepoCommands(context: ExtensionContext) {
     commands.registerCommand(
       `${EXTENSION_NAME}.copyRepositoryUrl`,
       async (node: RepositoryNode) => {
-        const url = getGitHubUrl(node.repo.name);
+        const url = getGitHubUrl(node.repo.name, node.repo.branch);
         env.clipboard.writeText(url);
       }
     )
@@ -99,7 +115,11 @@ export async function registerRepoCommands(context: ExtensionContext) {
     commands.registerCommand(
       `${EXTENSION_NAME}.copyRepositoryFileUrl`,
       async (node: RepositoryFileNode) => {
-        const url = getGitHubUrl(node.repo.name, node.file.path);
+        const url = getGitHubUrl(
+          node.repo.name,
+          node.repo.branch,
+          node.file.path
+        );
         env.clipboard.writeText(url);
       }
     )
@@ -109,7 +129,7 @@ export async function registerRepoCommands(context: ExtensionContext) {
     commands.registerCommand(
       `${EXTENSION_NAME}.openRepositoryInBrowser`,
       async (node: RepositoryNode) => {
-        const url = getGitHubUrl(node.repo.name);
+        const url = getGitHubUrl(node.repo.name, node.repo.branch);
         env.openExternal(Uri.parse(url));
       }
     )
@@ -119,7 +139,11 @@ export async function registerRepoCommands(context: ExtensionContext) {
     commands.registerCommand(
       `${EXTENSION_NAME}.openRepositoryFileInBrowser`,
       async (node: RepositoryFileNode) => {
-        const url = getGitHubUrl(node.repo.name, node.file.path);
+        const url = getGitHubUrl(
+          node.repo.name,
+          node.repo.branch,
+          node.file.path
+        );
         env.openExternal(Uri.parse(url));
       }
     )
@@ -141,11 +165,10 @@ export async function registerRepoCommands(context: ExtensionContext) {
               : path;
 
           await withProgress("Adding new file...", () =>
-            addRepoFile(node.repo.name, filePath)
+            addRepoFile(node.repo.name, node.repo.branch, filePath)
           );
 
-          const uri = Uri.parse(`repo:/${node.repo.name}/${filePath}`);
-          window.showTextDocument(uri);
+          openRepoDocument(node.repo.name, filePath);
         }
       }
     )
@@ -175,11 +198,14 @@ export async function registerRepoCommands(context: ExtensionContext) {
 
           return addRepoFile(
             node.repo.name,
+            node.repo.branch,
             filePath,
             // @ts-ignore
             contents.toString("base64")
           );
         });
+
+        openRepoDocument(node.repo.name, filePath);
       }
     )
   );
@@ -188,6 +214,7 @@ export async function registerRepoCommands(context: ExtensionContext) {
     commands.registerCommand(
       `${EXTENSION_NAME}.cloneManagedRepository`,
       async (node: RepositoryNode) => {
+        // TODO: Add support for branches
         const url = `https://github.com/${node.repo.name}.git`;
         commands.executeCommand("git.clone", url);
       }
@@ -204,9 +231,14 @@ export async function registerRepoCommands(context: ExtensionContext) {
             "Delete"
           )
         ) {
-          return withProgress("Deleting file...", () =>
-            deleteRepoFile(node.file)
-          );
+          return withProgress("Deleting file...", async () => {
+            const uri = RepoFileSystemProvider.getFileUri(
+              node.repo.name,
+              node.file.path
+            );
+
+            return workspace.fs.delete(uri);
+          });
         }
       }
     )
@@ -222,9 +254,18 @@ export async function registerRepoCommands(context: ExtensionContext) {
         });
 
         if (response && response !== node.file.path) {
-          await withProgress("Renaming file...", () =>
-            renameFile(node.file, response)
-          );
+          await withProgress("Renaming file...", async () => {
+            const fileUri = RepoFileSystemProvider.getFileUri(
+              node.repo.name,
+              node.file.path
+            );
+            const newFileUri = RepoFileSystemProvider.getFileUri(
+              node.repo.name,
+              response
+            );
+
+            return workspace.fs.rename(fileUri, newFileUri);
+          });
         }
       }
     )
@@ -234,6 +275,105 @@ export async function registerRepoCommands(context: ExtensionContext) {
     commands.registerCommand(
       `${EXTENSION_NAME}.refreshRepositories`,
       refreshRepositories
+    )
+  );
+
+  context.subscriptions.push(
+    commands.registerCommand(
+      `${EXTENSION_NAME}.createRepositoryBranch`,
+      async (node: RepositoryNode) => {
+        const user = await getCurrentUser();
+        const date = moment().format("L");
+        const branch = await window.showInputBox({
+          prompt: "Specify the name of the branch",
+          value: `${user}/${date}`
+        });
+
+        if (branch) {
+          await withProgress("Creating branch...", async () => {
+            await createBranch(node.repo.name, branch, node.repo.latestCommit);
+
+            await unmanageRepo(node.repo.name, node.repo.branch);
+            await manageRepo(`${node.repo.name}#${branch}`);
+          });
+        }
+      }
+    )
+  );
+
+  context.subscriptions.push(
+    commands.registerCommand(
+      `${EXTENSION_NAME}.deleteRepositoryBranch`,
+      async (node: RepositoryNode) => {
+        if (
+          await window.showInformationMessage(
+            `Are you sure you want to delete the "${node.repo.branch}" branch?`,
+            "Delete branch"
+          )
+        ) {
+          await withProgress("Deleting branch...", async () => {
+            await deleteBranch(node.repo.name, node.repo.branch);
+
+            await unmanageRepo(node.repo.name, node.repo.branch);
+            await manageRepo(`${node.repo.name}`);
+          });
+        }
+      }
+    )
+  );
+
+  context.subscriptions.push(
+    commands.registerCommand(
+      `${EXTENSION_NAME}.mergeRepositoryBranch`,
+      async (node: RepositoryNode) => {
+        const response = await window.showInputBox({
+          prompt: "Specify a description of your changes",
+          value: `Merging ${node.repo.branch}`
+        });
+
+        await withProgress("Merging branch...", async () => {
+          await rebaseBranch(node.repo.name, node.repo.branch, response);
+
+          await unmanageRepo(node.repo.name, node.repo.branch);
+          await manageRepo(node.repo.name);
+        });
+      }
+    )
+  );
+
+  context.subscriptions.push(
+    commands.registerCommand(
+      `${EXTENSION_NAME}.deleteRepository`,
+      async (node: RepositoryNode) => {
+        if (!globalStore.canDeleteRepos) {
+          return window.showErrorMessage(
+            'The token you used to login doesn\'t include the "delete_repo" scope.'
+          );
+        }
+
+        if (
+          await window.showInformationMessage(
+            `Are you sure you want to delete the "${node.repo.name}" repository?`,
+            "Delete repository"
+          )
+        ) {
+          try {
+            await withProgress("Deleting repository...", async () => {
+              await deleteRepository(node.repo.name);
+              await unmanageRepo(node.repo.name, node.repo.branch);
+            });
+          } catch (e) {
+            if (
+              await window.showErrorMessage(
+                "You don't have permission to delete this repository. Would you like to stop managing it?",
+                "Stop managing"
+              )
+            ) {
+              await unmanageRepo(node.repo.name, node.repo.branch);
+            }
+          }
+        }
+      }
     )
   );
 }